requirement.txt
Django==3.1.7
djangorestframework==3.12.2
pyroute2==0.5.14
python-iptables==1.0.0
pymemcache==3.4.2
Pillow==8.2.0
ipsetpy==0.0.1a3
django-cors-headers==3.7.0